Renaming FieldCheckboxMutliple to FieldCheckBox
